Use named variables
Programming best practice is to avoid using bare numbers, and instead to use variables – that way it is clear where values come from and their relationship to the underlying objects.  Where values are changed, it can be easy to miss a hard-coded number (as 1728 here), or to calculate its correct value by hand.